Rename babel-plugin -> babel-transform
